agent/requirements.txt
anthropic==0.42.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-asyncio==0.23.0
pytz==2024.1

# Download feature dependencies
playwright==1.40.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.105.0
pypdf==4.0.1
flask==3.0.0
